requirements.txt
openai
praw
jinja2
weasyprint
requests
python-dotenv
